# frontend/packages/types/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatbotAnalytics": {
        "properties": {
          "chatbot_id": {
            "format": "uuid",
            "title": "Chatbot Id",
            "type": "string"
          },
          "conversations": {
            "title": "Conversations",
            "type": "integer"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityPoint"
            },
            "title": "Daily",
            "type": "array"
          },
          "documents": {
            "$ref": "#/components/schemas/DocumentTotals"
          },
          "messages": {
            "$ref": "#/components/schemas/MessageTotals"
          },
          "window_days": {
            "description": "Length of the daily series, ending today (UTC)",
            "title": "Window Days",
            "type": "integer"
          }
        },
        "required": [
          "chatbot_id",
          "window_days",
          "documents",
          "conversations",
          "messages",
          "daily"
        ],
        "title": "ChatbotAnalytics",
        "type": "object"
      },
      "ChatbotCreate": {
        "properties": {
          "allowed_origins": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Allowed Origins",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "model_config_json": {
            "$ref": "#/components/schemas/GenerationConfig"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "privacy_url": {
            "default": "",
            "description": "Absolute URL of the tenant's privacy policy, shown in the widget footer. Empty string for no link.",
            "maxLength": 500,
            "title": "Privacy Url",
            "type": "string"
          },
          "retention_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Days of visitor conversation history to keep, counted from a conversation's last activity. Null keeps transcripts indefinitely, which is the default. A conversation held open by an unresolved ticket is never purged.",
            "title": "Retention Days"
          },
          "system_prompt": {
            "default": "",
            "maxLength": 8000,
            "title": "System Prompt",
            "type": "string"
          },
          "terms_url": {
            "default": "",
            "description": "Absolute URL of the tenant's terms, shown in the widget footer. Empty string for no link.",
            "maxLength": 500,
            "title": "Terms Url",
            "type": "string"
          },
          "theme_json": {
            "$ref": "#/components/schemas/WidgetTheme"
          }
        },
        "required": [
          "name"
        ],
        "title": "ChatbotCreate",
        "type": "object"
      },
      "ChatbotCreateResponse": {
        "properties": {
          "chatbot": {
            "$ref": "#/components/schemas/ChatbotRead"
          },
          "secret": {
            "$ref": "#/components/schemas/ChatbotSecret"
          }
        },
        "required": [
          "chatbot",
          "secret"
        ],
        "title": "ChatbotCreateResponse",
        "type": "object"
      },
      "ChatbotRead": {
        "properties": {
          "allowed_origins": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Origins",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_config_json": {
            "additionalProperties": true,
            "title": "Model Config Json",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "privacy_url": {
            "title": "Privacy Url",
            "type": "string"
          },
          "public_key": {
            "title": "Public Key",
            "type": "string"
          },
          "retention_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ChatbotStatus"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "terms_url": {
            "title": "Terms Url",
            "type": "string"
          },
          "theme_json": {
            "additionalProperties": true,
            "title": "Theme Json",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "name",
          "slug",
          "description",
          "system_prompt",
          "allowed_origins",
          "model_config_json",
          "theme_json",
          "retention_days",
          "privacy_url",
          "terms_url",
          "public_key",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ChatbotRead",
        "type": "object"
      },
      "ChatbotSecret": {
        "description": "The plaintext secret is returned exactly once, at creation or rotation.",
        "properties": {
          "chatbot_id": {
            "format": "uuid",
            "title": "Chatbot Id",
            "type": "string"
          },
          "secret_key": {
            "title": "Secret Key",
            "type": "string"
          }
        },
        "required": [
          "chatbot_id",
          "secret_key"
        ],
        "title": "ChatbotSecret",
        "type": "object"
      },
      "ChatbotStatus": {
        "enum": [
          "active",
          "paused",
          "archived"
        ],
        "title": "ChatbotStatus",
        "type": "string"
      },
      "ChatbotUpdate": {
        "properties": {
          "allowed_origins": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Origins"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "model_config_json": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "privacy_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Url"
          },
          "retention_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Days of visitor conversation history to keep, counted from a conversation's last activity. Null keeps transcripts indefinitely, which is the default. A conversation held open by an unresolved ticket is never purged.",
            "title": "Retention Days"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatbotStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "system_prompt": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "terms_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Url"
          },
          "theme_json": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WidgetTheme"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ChatbotUpdate",
        "type": "object"
      },
...
      "EmbedSnippet": {
        "properties": {
          "loader_url": {
            "title": "Loader Url",
            "type": "string"
          },
          "public_key": {
            "title": "Public Key",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          }
        },
        "required": [
          "public_key",
          "loader_url",
          "snippet"
        ],
        "title": "EmbedSnippet",
        "type": "object"
      },
...
      "GenerationConfig": {
        "properties": {
          "max_tokens": {
            "default": 1024,
            "maximum": 8192.0,
            "minimum": 64.0,
            "title": "Max Tokens",
            "type": "integer"
          },
          "min_similarity": {
            "default": 0.25,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Similarity",
            "type": "number"
          },
          "temperature": {
            "default": 0.2,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          },
          "top_k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "title": "GenerationConfig",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ChatbotRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatbotRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ChatbotRead]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WidgetChatRequest": {
        "properties": {
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_\\-]+$",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "message",
          "session_id"
        ],
        "title": "WidgetChatRequest",
        "type": "object"
      },
...
      "WidgetTheme": {
        "additionalProperties": false,
        "description": "How the widget looks on the tenant's site.\n\nEvery member is optional and only what was set is persisted. The widget applies these as\ninline custom properties over its own stylesheet, so an unset colour keeps the default \u2014\nincluding the automatic dark-mode switch \u2014 while a set one wins outright.\n\nSix digits of hex, no shorthand and no `rgb()`: the values are interpolated into a style\nattribute, and a narrow pattern is easier to be sure of than an escaping routine.",
        "properties": {
          "accent": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent"
          },
          "accent_foreground": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent Foreground"
          },
          "border": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Border"
          },
          "greeting": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Greeting"
          },
          "position": {
            "anyOf": [
              {
                "enum": [
                  "right",
                  "left"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "radius": {
            "anyOf": [
              {
                "maximum": 28.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius"
          },
          "scheme": {
            "anyOf": [
              {
                "enum": [
                  "system",
                  "light",
                  "dark"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheme"
          },
          "surface": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "surface_muted": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Muted"
          },
          "text": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "text_muted": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Muted"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "WidgetTheme",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Dashboard access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RAG Chatbot Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chatbots": {
      "get": {
        "operationId": "list_chatbots_api_v1_chatbots_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChatbotStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ChatbotRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Chatbots",
        "tags": [
          "chatbots"
        ]
      },
      "post": {
        "operationId": "create_chatbot_api_v1_chatbots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Chatbot",
        "tags": [
          "chatbots"
        ]
      }
    },
    "/api/v1/chatbots/{chatbot_id}": {
      "delete": {
        "operationId": "delete_chatbot_api_v1_chatbots__chatbot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Chatbot",
        "tags": [
          "chatbots"
        ]
      },
      "get": {
        "operationId": "get_chatbot_api_v1_chatbots__chatbot_id__get",
        "parameters": [
          {
            "description": "Chatbot identifier",
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "description": "Chatbot identifier",
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Chatbot",
        "tags": [
          "chatbots"
        ]
      },
      "patch": {
        "operationId": "update_chatbot_api_v1_chatbots__chatbot_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Chatbot",
        "tags": [
          "chatbots"
        ]
      }
    },
...
    "/api/v1/chatbots/{chatbot_id}/analytics": {
      "get": {
        "description": "Ingestion and conversation counters for the dashboard overview.",
        "operationId": "chatbot_analytics_api_v1_chatbots__chatbot_id__analytics_get",
        "parameters": [
          {
            "description": "Chatbot identifier",
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "description": "Chatbot identifier",
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotAnalytics"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chatbot Analytics",
        "tags": [
          "chatbots"
        ]
      }
    },
...
    "/api/v1/chatbots/{chatbot_id}/embed-snippet": {
      "get": {
        "operationId": "embed_snippet_api_v1_chatbots__chatbot_id__embed_snippet_get",
        "parameters": [
          {
            "description": "Chatbot identifier",
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "description": "Chatbot identifier",
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedSnippet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Embed Snippet",
        "tags": [
          "chatbots"
        ]
      }
    },
    "/api/v1/chatbots/{chatbot_id}/rotate-secret": {
      "post": {
        "operationId": "rotate_secret_api_v1_chatbots__chatbot_id__rotate_secret_post",
        "parameters": [
          {
            "in": "path",
            "name": "chatbot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chatbot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotSecret"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rotate Secret",
        "tags": [
          "chatbots"
        ]
      }
    },
...
    "/public/widget/chat": {
      "post": {
        "description": "Stream one answer over SSE.\n\nOnce the first byte is on the wire the status code is fixed, so failures after that\npoint are delivered as an `error` event and the stream is closed cleanly.",
        "operationId": "chat_public_widget_chat_post",
        "parameters": [
          {
            "description": "Public widget key (pk_...)",
            "in": "header",
            "name": "X-Chatbot-Key",
            "required": true,
            "schema": {
              "description": "Public widget key (pk_...)",
              "title": "X-Chatbot-Key",
              "type": "string"
            }
          },
          {
            "description": "Origin of the page hosting the widget, as attested by the browser",
            "in": "header",
            "name": "X-Widget-Site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Origin of the page hosting the widget, as attested by the browser",
              "title": "X-Widget-Site"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WidgetChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "widget"
        ]
      }
    },
...
  }
}